Crate defmt_parser

Parsing library for defmt format strings.

This is an implementation detail of defmt and not meant to be consumed by other tools at the moment so all the API is unstable.

Structs

Parameter

A parameter of the form {{0=Type:hint}} in a format string.

Enums

DisplayHint

All display hints

Error

The kinds of error this library can return

Fragment

A part of a format string.

Level

The log level

ParserMode

Parser mode

TimePrecision

Precision of timestamp

Type

Functions

get_max_bitfield_range

Returns Some(smallest_bit_index, largest_bit_index) contained in params if params contains any bitfields. Otherwise None.

parse